# expression: implement vectorized evaluation for `builtinArithmeticMultiplyRealSig`