### What problem does this PR solve?

Implement vectorized evaluation for builtinArithmeticMultiplyRealSig.#12105

### What is changed and how it works?
```
goos: linux
goarch: amd64
pkg: github.com/pingcap/tidb/expression
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticMultiplyRealSig-VecBuiltinFunc-4                384745              3277 ns/op            0 B/op           0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticMultiplyRealSig-NonVecBuiltinFunc-4              39608             29081 ns/op            0 B/op           0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusRealSig-VecBuiltinFunc-4                    270891              3777 ns/op            0 B/op           0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticPlusRealSig-NonVecBuiltinFunc-4                  40618             28557 ns/op            0 B/op           0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticMinusRealSig-VecBuiltinFunc-4                   299098              3472 ns/op            0 B/op           0 allocs/op
BenchmarkVectorizedBuiltinArithmeticFunc/builtinArithmeticMinusRealSig-NonVecBuiltinFunc-4                 37515             31396 ns/op            0 B/op           0 allocs/op
PASS
ok      github.com/pingcap/tidb/expression      8.894s
```

### Check List

Tests
- Unit test